Extract BerthOfferTable
